requirements.txt
arcade==2.6.17
python-dotenv==1.0.1